requirements.txt
matplotlib
seaborn
plotly
streamlit_option_menu